requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
djangorestframework==3.14.0
python-dotenv==0.21.0
python-environ==0.4.54
pytz==2022.7
sqlparse==0.4.3
psycopg2-binary==2.9.3
django-cors-headers==3.13.0
requests==2.28.0